+++ python/ray/serve/_private/default_impl.py
@@ -29,6 +29,7 @@ from ray.serve._private.utils import (
     inside_ray_client_context,
     resolve_request_args,
 )
+from ray.serve.handle_options import _DynamicHandleOptions, _InitHandleOptions
 
 # NOTE: Please read carefully before changing!
 #
@@ -56,14 +57,10 @@ def create_deployment_scheduler(
 
 
 def create_dynamic_handle_options(**kwargs):
-    from ray.serve.handle import _DynamicHandleOptions
-
     return _DynamicHandleOptions(**kwargs)
 
 
 def create_init_handle_options(**kwargs):
-    from ray.serve.handle import _InitHandleOptions
-
     return _InitHandleOptions.create(**kwargs)
